feat(setcodetx): fix funding and confirmation handling in max bloating mode
- Updated DefaultTargetGasRatio to 99.5% for lowered safety margin.
- Modified fundMaxBloatingDelegators to return the confirmed count of funded transactions.
- Implemented waitForFundingConfirmations to ensure funding transactions are confirmed before proceeding.

Even with all that, it's inevitable to get bloating phases split across blocks if it happens that they are sent when 1 block time is almost going to be terminated.
Nevertheless this is better than leaving a full block empty to fit them both in the next one. So the behaviour will persist.
Also, this is logged so that the user knows this.